Logging output for tests in file
This commits redirects all logging output for unit tests into a file in
the workspace. This is achieved by specifying a logging configuration
for the test runs and by letting every test class extend from
LoggingEnabled, which includes a JUnit Rule and Watcher which passes the
test names to the logging system. Every new test needs to extend this
class!